Create Shape
A program in cpp create an Abstract class called 'Shape' to model a 3-D shape. We derive classes Cube and Cone from the shape class. The user will specify the size of each shape. The user may specify many item of the same shape. All the instances of each shape should be stored in a single
array. The program will then calculate area and volume of each shape.